Decoupling socket server and handler
Segregation of concerns - BluetoothServer is responsible for waiting for connections and then delegating to the SocketHandler
